requirements.txt
django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-environ==0.11.2
psycopg2-binary==2.9.9
Pillow==10.3.0
django-filter==24.2